change integer type of window

+++ Hyades/src/Window.hpp
@@ -11,24 +11,24 @@ namespace Hyades
 
         GLFWwindow* m_window;
 
-        int m_width{ 800 };
-        int m_height{ 600 };
+        size_t m_width{ 800 };
+        size_t m_height{ 600 };
 
         const std::string m_title{ "Hyades Window" };
 
     public:
-        Window(const std::string& title, const int& width, const int& height);
+        Window(const std::string& title, const size_t& width, const size_t& height);
         ~Window();
 
         void on_update();
         
         // getters
-        int const& width() const { return m_width; };
-        int const& height() const { return m_height; }; 
+        size_t const& width() const { return m_width; };
+        size_t const& height() const { return m_height; }; 
 
         // setters
-        int& width() { return m_width; }
-        int& height() { return m_height; }
+        size_t& width() { return m_width; }
+        size_t& height() { return m_height; }
 
         // callbacks
         static void set_window_size_callback(GLFWwindow* window, int width, int height);
